src/requirements.txt
cryptography==44.0.1
Flask==3.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.12
Flask-Cors==5.0.0
loguru==0.7.2
numpy==1.26.4
onnxruntime>=1.8.0
opencv-python>=4.5.0
opencv-python-headless==4.10.0.84
Pillow==10.3.0
requests==2.32.3
mcp==1.13.1
torch==2.2.2
torchvision==0.17.2
